random number guessing game
this is a game in which sets for a random number each time the game is played. The game will let the user know if their guess is to high or to low. Say that the user had got the number right on the dot then the game will then tell the user so and quit.